Minor cleanups

+++ src/common.h
@@ -118,7 +118,7 @@ typedef struct _IdleData {
     gboolean showcontrols;
     gdouble position;
     gdouble length;
-    gint byte_pos;
+    glong byte_pos;
     gint brightness;
     gint contrast;
     gint gamma;
